pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ickyMCP"
version = "0.1.0"
description = "RAG MCP Server for Document Search - supports PDF, Word, Excel, PowerPoint"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Devansh"}
]
keywords = ["mcp", "rag", "embeddings", "document-search", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "sqlite-vec>=0.1.0",
    "tiktoken>=0.5.0",
    "pypdf>=3.0.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.0",
    "openpyxl>=3.1.0",
    "pathspec>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
ickymcp = "src.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
